Fix memory leak in frame varlist reuse
This is a clumsy though reasonably robust way of dealing with a
situation that was leading to not noticing when a frame was pushed,
received a varlist via reuse, and then overwrote that varlist with its
own data.  It means that Push_Frame_Core() no longer provides a varlist
reuse by default, it must be explicitly asked for, making it more
noticeable at callsites whether they explicit ask for it or not...
and if they do not ask but actually wanted one, it will crash due to
the pointer being trash.

It can probably be done more elegantly, but this resolves a leak that
was observed in Travis during make.r under address sanitizer.