Rename "Apply hook" to "Dispatcher hook"
In examining the C stack, it was confusing to see Apply_Core, as one
might think APPLY was running--or something related to it.  Really
what it's doing is allowing one to hook ACTION! dispatchers, so this
goes ahead and renames it--as well as giving the trace/measured hooks
more palatable names.