Abandon "ghost" pointer experiment due to lack of benefit
Instead of using null for UNBOUND, a special global REBSER struct was
experimented with.  It was at a location in memory known at compile
time, and it had its ->header and ->info bits set in such a way as to
avoid the need for some conditional checks.  e.g. instead of writing:

    if (binding and binding->header.bits & NODE_FLAG_MANAGED) {...}

The special UNBOUND node set some bits, e.g. to pretend to be managed:

    if (binding->header.bits & NODE_FLAG_MANAGED) {...} // incl UNBOUND

Question was whether avoiding the branching involved from the extra
test for null would be worth it for consistent ability to dereference.
At least on x86/x64, the answer was: No.  It was maybe even a little
slower--though seemed to be within the noise.

Testing for null pointers the processor has in its hand is very common
and seemed to outweigh the need to dereference all the time.  The
increased clarity of having unbound be nullptr is also in its benefit.

Also changes a LIT-BAR! in a test to the word BAR because for mystery
reasons (currently being investigated) that causes some kind of skew
that makes the tests take twice as long to run.  While that's being
investigated, this changes it--though it really is a top priority
to figure out how that could possibly be having that effect.