Allow arbitrary REBVAL* to be returned from dispatchers
R3-Alpha's native implementations could return an enumerated type to
describe where their return value could be found--either on the stack,
or in the list of input arguments.  Ren-C rethought that and added
more to the enumeration beyond just where to find values.  This
included instructions that were used by "dispatchers"--which had
native interfaces, but were for processing sub-dispatch or implementing
function compositions.

This commit changes the interface to no longer return an enumerated
type, but to return a pointer which could be to a value -or- to a
special non-value whose first byte can be switched on as if it were
an enumerated type.  How it works is by taking advantage of the fact
that REBVAL* and char* for UTF-8 strings don't overlap--which means
that so long as the first byte is < 127 it can be used as if it
were an enumerated type.