Fix HIJACK of a specialization
The mechanism for translating a frame build for a function to the
function that has hijacked it wasn't accounting for hidden parameters
in a specialization frame.  Hence the hijacked routine would expect
all the underlying function's arguments to be provided--even if at
the user level it appeared they were hijacking a function with fewer
apparent arguments after the specialization.