Eliminate R_OUT, R_OUT_IS_THROWN, R_NULL
The "R_OUT_IS_THROWN" state is something that was tending to be thought
of as unnecessary.  The system can only have one throw state in effect,
and having the user annotate when they were throwing was just something
that was checked in the debug build to tell you that it was mismatched.

What emerged as the main solution to throws being dropped was the
idea of having the APIs that can throw return a flag.  Then, avoid
having an API for "thrown" detection altogether, which forces the
clients of the evaluator to come up with a protocol for conveying the
thrown state.

That plan isn't fully finished, but most all usages of THROWN are
currently in the debug build...and the flag state itself will likely
not be something that gets put on cells at all (except, again,
possibly in the debug build).  This goes ahead and eliminates the
distinction between R_OUT_IS_THROWN and R_OUT, and then just uses
D_OUT instead of R_OUT as the return value.

It also eliminates R_NULL, as the use of nullptr is more elegant, and
is needed for being able to return any API routine result from rebRun()
into a native return value.