bound? => chain of `binding of` into :value?
Historical Rebol had two synonyms for binding: BOUND? and BIND?.
Though they ended in `?`, they did not return LOGIC!, rather the
context a word was bound to or NONE!.

Ren-C is getting rid of operations that end in `?` but do not return
a LOGIC!.  The reflector BINDING used with OF was chosen to produce
the syntax `binding of word`, to return either the context a word is
bound to or NULL.

The BOUND? and BIND? functions were deprecated.  However, BOUND? is
actually a good name for a LOGIC!-returning test of whether the
BINDING OF is equal to NULL or not, and it was in use in the console.

This redefines BOUND? and removes the deprecation error from its use,
while leaving the error for BIND?