Further formalize invariants for bindings, fixes bug
As the various forms of binding take shape, one issue has been how it
would be possible to only pay the cost for "reification" of stack
structures if their references would leak into cells that would outlive
those stack structures.

The tool for being able to make such decisions is the existence of
Move_Value() and Derelativize() as the gatekeepers for moving values
from one location to another.  The rules have been emerging over time,
and this commit adds some asserts that enforce the rules, and conforms
with those asserts.

One rule that was being broken in the previous commit (that was causing
a build bug on OS X Travis) was that stack frame contexts could not
become managed unless the action frame had completed its fulfillment
process.  The implementation of APPLY was allowing this, because it
was using the same frame context references for the variables in the
block as it used when it executed the frame.  For now, this works
around it by "stealing" the context's variables, so the APPLY's block
can't be used as a back door to accessing an in-progress frame once
the function begins running.