Remove micro-optimizations from ELSE, ALSO
The savings of Move_Value() prior to returning are now rather small in
the scheme of things.  All values are test for THROWN(), even D_OUT.
So what it would do is avoid a null pointer test and a flag test.

At this point, it's better for the fundamental routines to "show well"
and have people read and understand them, as opposed to worry about
something quite that small.

Also in the interest of furthering understanding, this changes the
order of the arguments to Run_Branch_Throws to put the branch first
and the parameter to it second.