Retake THEN as complement to ELSE, give ALSO "SO"'s behavior
The "obvious" concept that THEN and ELSE are related as fairly precise
antonyms/complements led them to be defined that way initially.  Yet
mechanics for simple cases like `all [...] then [...]` did not work
out, due to routines like ALL returning a conditionally false result
but that still was an ANY-VALUE!.  This meant that THEN would always
run, if it were testing the left hand side for "ANY-VALUE!" vs.
testing for conditional truth.

That led THEN to be defined as being just an infix form of IF.  Yet
the mechanic of constructs like ANY, ALL, FIND, PARSE, etc. have
changed to give no-result back as NULL.  That means the original issue
that was being worked around is no longer applicable.

This deletes THEN and NOT-THEN under their old meanings, giving THEN
the meaning that ALSO had, and moving ALSO to the behavior which was
being called "SO".  This shuffle didn't affect that much code, since
most usages of THEN were of the form `all [...] then [...]`, and that
still works.

Also fixes an assert that was tripped on during testing, relating to
the optimization passthru for building action frames without calling
the action.