GC guard synthetic paramlist during APPLY
An APPLY is much like running a SPECIALIZE'd action--and shares much
of the same code.  However, a unique ACTION! with the specialized
parameters hidden is not created.  The only ACTION! available is the
original one.

So the "facade" to use is really the keylist of the specialized
temporary context.  The frame is given that context's params as its
f->param to start out with.  However, since the context is temporary,
it could be GC'd...leaving the f->param fields free.

This extends the GC guard past just the DO of the apply definition code
into the temporary frame, to last for the duration of the entire
action execution.