Cut down SPECIALIZE/APPLY memory, less redundant typechecks
Previously, when specializations were made they would have to create
three new arrays.  One array would be the same length as the parameter
list, and hold the specialized values (null in any unspecialized slots)
Another would be a new identity parameter list with fewer arguments,
each specialized parameter removed.

The third array was a "facade"--a list of typesets just as long as the
original paramlist, but which would have a bit set on any removed
parameters.  This would be used in filling the frame of the appropriate
size to use with the real action--because it had the right number of
parameters, and the information of which to skip.

This commit gets rid of the 3rd array, by sneaking an "out of band" bit
into the specialized argument values.  The bit is not specific to
specialization, however.  It more generically indicates that a value
has been typechecked.  When value cells get overwritten it will disrupt
this flag, cueing processes not to typecheck it again.

APPLY benefits proportionally more, cutting it down to creating just
one new array instead of two (which due to previous work, gets actually
spliced in as the stack variables for the invocation.)

There are lots of advantages to establishing the new bit as a generic
"already type checked" bit vs having a bit just for specialization.
It reduces the amount of typechecking needed on things like ADAPT or
REDO.  Only arguments that have been modified will be re-checked, which
may be a longer process once there are more complex typeclasses and
user-defined types.

Another benefit is it removes the need for the convoluted code which
the previous commit's patch had to be added to fix.  Now a phase's
paramlist is always used, not needing to patch in "stolen" context
keys as the parameters...since the information it wants is resident
in the varlist.

It's a significant change to the mechanics established during the
addition of partial refinement specialization.  But for the most part,
it's a simplifying change--which brings it closer to being able to
do things like MAKE ACTION! directly from a FRAME! value, allowing
users to easily create their own SPECIALIZE-like actions.