Fix hole in GC model, simplify GC guards, MATCH GC bugfixes
R3-Alpha considered all series to be managed.  If code ran that could
trigger the evaluator or otherwise run a GC, then any series that had
been created and were linked to by C local variables--but but were not
reachable from the root set--needed to have PUSH_GUARD() on them,
else they would be candidates for getting freed.

Ren-C introduced the idea of a "managed" bit on series.  Any series
without it would not be a candidate for freeing.  Putting a series
into a REBVAL cell would automatically manage it, though it could be
handled other ways.  Manually freeing a series was then only allowed
if it was in an unmanaged state.

One thing that was not addressed was the GC protection of the contents
of manual arrays.  If code was building up an array one element at
a time, and managed to trigger a GC before the array was finished, then
the unmanaged array's reference in its cells would not be enough to
keep things alive.  This necessitated PUSH_GUARD_ARRAY_CONTENTS, which
was a way of saying that although the array wasn't managed (and hence
not a candidate for PUSH_GUARD_ARRAY, but a candidate for being freed),
its contents needed protection.

But if manual arrays are to exist at all across a GC, then *all*
manual arrays woul dneed PUSH_GUARD_ARRAY_CONTENTS.  Hence this commit
eliminates that routine, and switches it so that manual series are
treated as "roots" in the GC set, and will hold their contents alive.

The implication is that anything which builds a manual array must fill
its claimed length with GC-valid content (even if that content is
UNREADABLE_BLANK) to survive across a GC.  There's another issue
involved regarding how the LINK() and MISC() fields are handled, but
this implements a first cut of the rule to see how it works

Additionally, the ability to push guards for contexts and values--
as well as plain series and arrays--had led to a unification of the
guarding process to be for "nodes".  At the time of that unification
it was speculative, and so typed routines like PUSH_GUARD_CONTEXT and
PUSH_GUARD_ARRAY existed.  This unifies it to PUSH_GC_GUARD() and
DROP_GC_GUARD(), which make what's being guarded clearer and can act
generically across all node types.

Includes bugfixes to MATCH noticed while testing.