Make DEFAULT allow `<end>` on left, native implementation
This enables DEFAULT to be used with CASE and SWITCH to give a
lexically clean way of defaulting:

    x: 10
    case [
        x = 20 [print "this won't run"]
        default [print "yay, normal-looking default!"]
    ]

    switch 'some-word [
        'default [print "would match the *word* DEFAULT"]
        default [print "yay, normal-looking default!"]
    ]

It does this without damaging DEFAULT's existing usages when it sees
a SET-WORD! or a SET-PATH! on its left, to get that word or path and
check to see it if is blank or NULL and only override it then:

    >> x: _
    >> x: default [10]
    >> print x
    10
    >> x: default [20]
    >> print x
    20

The reason the new usage is possible is because SWITCH now evaluates
its clauses, and that both CASE and SWITCH have a "fallout" feature
of evaluating single clauses.  So it's not that DEFAULT is matching
in particular, it's just single-arity and there's no matching branch:

    case [
       (condition) [...branch...]
       (condition) [...branch...]
       (condition) ;-- reach END, no branch, condition value falls out
    ]

So `DEFAULT [...]` is not a condition and a branch, it's just a
condition.  It enforces that it can't be used with a branch, by making
sure if it has no left hand side that it has no data after the block.

While it's something of a "trick", it definitely raises the legibility
and helps to document the intent.  With CASE, people have traditionally
accomplished this by making a condition TRUE as the last case, with
debatably more comprehensible choices like /ELSE which are also
truthy.  But that would not work for SWITCH, while this will, and this
gives a good legible standard...while not ruling out those uses for
people who insist on using them with CASE.

Fortunately, this can be made to work in the shim for older Ren-Cs, so
it allows cleaning up some of the less-legible defaulting in %rebmake.r

Since this means DEFAULT will be used even more frequently than it
already is, this commit goes ahead and makes it a native, so it will
be faster.