Fix GC interaction with ENCLOSE
The Encloser_Dispatcher "steals" variables from the REBFRM's varlist,
which may-or-may not be managed at the time of the "theft".  If it
was managed, the garbage collector was set up to recognize the
SERIES_INFO_INACCESSIBLE state which said it no longer had variables.
But the unmanaged state depended on the frame stack to mark variables,
which didn't recognize the state.  This adds a check for that.

Found via long-running process of R3_RECYCLE_TORTURE on all the tests.