Always have one stack frame in effect
In order to avoid conditional checking for null of the previous frame,
this makes sure there is at least one "dummy" frame in effect at the
bottom of the stack.  It also consolidates the dummy ACTION! used for
that frame's phase with two other instances in which a placeholder
action was required in the code.

Truly being able to use the top frame in all operations and know it
was not null...e.g. in API operations, required some reodering of the
startup process.  This process needs review--especially in light of
threading or isolates.  However, minimal changes were made just to get
the necessary dependencies for this commit.

One result of this is that there is always a context to which to tie
Alloc_Value() allocations.