Streamline native and port actor return values
A recent change allowed returning arbitrary value pointers from natives
so that things like:

    Move_Value(D_OUT, ARG(foo));
    return R_OUT;

...could be simplified to:

    return ARG(foo);

This opens up some possibility for improvement.  For instance, in the
port code, there were many cases where code would jump to a return_port
label, e..

    goto return_port;
    ...
  return_port:
    Move_Value(D_OUT, D_ARG(1));
    return R_OUT;

This doubly simplifies it, by making the `port` argument to the actor
a REBVAL instead of a REBCTX, so one can just say:

    return port;

That's more succinct at the callsite than `goto return_port;` while
retaining the clarity, and gets rid of the need for the label at all.

Also, the Init_XXX functions return the target location, so:

    Init_Integer(D_OUT, num);
    return D_OUT;

Can now be reduced to just:

    return Init_Integer(D_OUT, num);

While a small difference, it reduces the total number of lines of code,
and when one does that it becomes easier to see what's going o.