Fix behavior of EXISTS? for root directory %/
The TO-LOCAL-FILE function had the option to omit a trailing slash.
When %/ was expanded to a "full" path with the trailing slash removed,
this would yield an empty string.  Neither GetFileAttributesEx() on
Windows nor UNIX's stat() would work with an empty string.

Reasoning for why tail slashes were removed from directories for these
APIs are not clear, as things appear to work with them included.  Hence
the easiest workaround for the problem seems to be to leave the slash.