Fix MAKE BLOCK! of MAP!, misc fixes
The code for MAKE_Block would fall back on TO_Block if it didn't have
specific functionality.  However, the fallback was being run and
overwriting the specific cases (e.g. turning a MAP! into a block of
individual elements, vs. TO's behavior of putting the map into a
1-element block).

This fixes that and removes the TO fallback--as part of a general
push to clearly define what the difference between a MAKE and a TO
actually is.

Includes a couple of other bug fixes found while testing--such as
an unknown refinement reporting the error as an ISSUE!.