Make RETURN endable, all can be arity 0 -or- 1
This commit changes it so that RETURN can always be given an argument
or not.  If it is not given an argument, then it acts just the same
as if someone had written RETURN VOID.  Hence the following is legal:

     foo: func [x] [
         if x > 0 [return] else [return 100]
     ]

The return type will still be type checked, so void must be one of the
legal return types for the function.  (It is legal if there is no
return type annotation given.)

Historical Rebol had two different forms of RETURN...an arity-1 version
called RETURN, and an arity-0 version called EXIT.  Either form could
be used from any function, and invoking EXIT would mean that the
result would be an UNSET!.

Definitional returns changed the mechanics of RETURN significantly, and
there were two different kinds of function generators: FUNC/FUNCTION
which defined an arity-1 RETURN for each invocation, and PROC/PROCEDURE
which defined an arity-0 LEAVE for each invocation.  These were
eventually fused back to where "PROCEDURE" semantics were achieved by
just a FUNC or FUNCTION that said `return: <void>` in its spec, and
this affected whether the RETURN would be arity-0 or arity-1.