Use THEN/ELSE/ALSO/AND/OR in bootstrap, endable FAIL
It was mistakenly believed that the committed build for bootstrap had
left enfix normal behavior that was not completing, thus it would be
different from the current behavior.  However, the bootstrap build
is actually old enough that it can be more-or-less compatible with
current left normal enfix.

Hence this opens the doors to using THEN and ELSE, and also fills in
implementations of ALSO, AND, and OR.  Introduces a few usages in the
make process just to see them working.  Also goes ahead and kills off
the support for FORSKIP, FORALL, and FOREACH (FOR-SKIP, FOR-NEXT, and
FOR-EACH should be used instead).

Also makes it so that FAIL can be used with no arguments as a lazy
form of failure, which cites only the location but provides no unique
failure message.