Fix build breakage due to TO BLOCK! change
The behavior of TO BLOCK! for a PATH! had changed to put the path
into a BLOCK! (not to coerce it to a BLOCK! the way AS BLOCK! would).
This broke a build, but made it look like it timed out--which Travis
has been bad about doing lately.  However, it really was timing out
because tests were breaking into a console and just sitting there in
the second build.

The case that used it didn't really need to use TO BLOCK!; it was doing
a FIND that it could have done on a PATH!.