Rename %reb-lib.h to %rebol.h, API restructuring
The external API is the most sensible owner of the name "rebol.h", so
this renames it to take it (nothing was using it ATM).

Also, since the core uses the internal API, it should theoretically be
able to include the API header *before* any of the implementation
definitions.  This wasn't the way it had been done--instead, there
was conditional #ifdef code based on REB_DEF, that caused a file called
%reb-defs.h to act differently.  API clients had to include both
%reb-lib.h and %reb-defs.h to get the definition of REBVAL--which
was a void*, that isn't as type-safe as just using an opaque pointer.

This goes ahead and makes that change, so that %rebol.h is the only
file that needs to be included (besides %reb-c.h, which could be
substituted with #include'ing stdbool.h and stdint.h.)  Hence the
%reb-defs.h file is used only by the internal API, and could be
redefined as "the include file that gets Rebol to the state where
the internal API prototypes can be compiled".

In the process, this removes a couple of barriers to that factoring.
It eliminates rebEvent() the API, in favor of having code that wants
to raise an event that doesn't have internal API access to call into
the Rebol `make event!` code.  It also removes the exposure of the
REBXXX types (like REBINT) in the libRebol API, as using plain C
types makes more sense there.