Use bool instead of REBOOL in API, Blit_Cell vs. memcpy
Both C99 and C++ have a bool type.  There is no clear rule in the
standard that both would be the same size, however, consider:

"There doesn't seem to be any guarantee that C int is compatible with
 C++ int. 'Linkage from C++ to objects defined in other languages and
 to objects defined in C++ from other languages is implementation
 defined (...) I'd expect C and C++ compilers intended to be used
 together (such as gcc and g++) to make their bool and int types, among
 others, compatible."

https://stackoverflow.com/q/3529831

Given this rationale, it makes little sense to be abstracting REBOOL
in the API.  We can take for granted that `bool` would be compatible
in C99 with any version of C++, and if one is being forced to use a
C89 compiler then the constraints of the platform are probably such
that you're not compiling as C++ at all...much less mixing and matching
a C++ build with C built products.

Hence this goes ahead and uses `bool` in the libRebol API; but with
the small hedge that it is not assumed to be exactly 0 or 1 when
passed in by a caller, but clamped with DID.

It also corrects a usage of memcpy() on cells when collecting keylists
to use Blit_Cell().  (memcpy was being used for "performance" to avoid
a loop, but then a loop was getting run anyway to add binding indices
in a second phase, so it wasn't really doing any good anyway, but
memcpy() shouldn't be used on cells generally anymore).  Also tames
some excessive commenting.

Inserts a couple of API usages at fail() sites, since fail will now
clean up API handles from failed frames.