Eliminate VAL_TYPE_OR_0 "optimization", 2-byte rebEnd trick
In an effort to reduce the amount of conditional branching, the global
END marker was given a type of REB_0.  This allowed there to be
a special VAL_TYPE_OR_0 operation which--unlike the ordinary VAL_TYPE()
would permit use on an END.  Hence you could write things like:

    if (VAL_TYPE_OR_0(v) == REB_ACTION) {...}

instead of:

    if(v != END and VAL_TYPE(v) == REB_ACTION) {...}

This couldn't be used on arbitrary ENDs, however--since those could
use the type bits for any purpose if they were "Endlike headers".
So really this was a trick for when the compared value was using
END instead of null.  So it actually was trying to improve on:

    if (v and VAL_TYPE(v) == REB_ACTION) {...}

It turns out that in practice, this optimization offers no savings.
Tests for null are fast, and branch predictors appear to handle this
pretty well, and avoiding dereferences in the null case can be good.
Hence it was obfuscating for no real advantage, and also had a complex
rule to explain ("it's only the global END, not just any END"...)

A very concrete advantage of ditching this optimization is that the
globally used END does not need to be large enough to use its type
slots.  It can be short...in fact, with trickery, it can be just two
bytes long...the string literal "\x80".

This makes that change, with:

    #define rebEND \
       ((const void*)"\x80")

This makes the API use not require linkage to a global variable.