Allow arbitrary length "details", fix FFI GC interaction
Every ACTION! has at least two unique arrays.  One for the parameter
list (an array whose [0] slot is the archetypal ACTION! value, and
then 1...N TYPESET!-with-symbol parameters).  The other array has
whatever information is needed to inform the dispatcher C function
how to run that instance of that dispatcher's function class.

Previously the second array was always exactly 1 element long, as an
optimization, to make use of the "singular" array.  The idea was that
if an instantiation needed more than that, it could initialize that
one element with a BLOCK! or similar.  It was called a `body_holder`
and accessing the body always gave back a relative value, with
ACT_BODY().

Today the singular optimization is automatic, so asking for an array of
length 0 or 1 will get a singular.  So this makes a change such that
when Make_Action() is called, you pass a capacity of how long you want
the second array to be (now called the action's "details").

Since this can now be any size, the dispatchers that previously had to
allocate arrays (e.g. ADAPT, ENCLOSE) now just ask for "details" the
length they need.  Similarly, the FFI's routines and callback needed
several tracking values in an array known as a REBRIN ("Rebol Routine
Information"), which now avoids a level of indirection by being the
details array.

(Note: The name "info" was considered for this array, but would be
confusing since "info bits" are already taken, and relevant to actions,
so ACT_INFO() would be sort of ambiguous.)

While testing the FFI with the change under RECYCLE/TORTURE, it was
discovered that it broke the constraint that manually managed arrays
must contain GC-valid data (at minimum, unreadable blanks) if they are
to be held across evaluations.  So that is fixed as well.