rebJUMPS => rebJumps, remove "lib/" prefixes on API calls
This removes the lib prefix on API calls in anticipation of upcoming
changes to make the runtime sensitive to the native code that is
running when doing binding.

(It doesn't do it on the %host-main.c yet, because since it's at the
top level there is no native to attach it to...and the most sensible
default for API clients is to have the default be the user context.
So the lib/ stays there for now.)

At the same time, it takes the caps off of the rebJUMPS API, instead
favoring the capitalization of the operation which does the FAIL or
the PANIC.  This looks like a little less of a non-sequitur.