Associate each native with an ANY-CONTEXT! for binding
This implements a first cut of associating natives with contexts, so
that while that native is running the rebXXX() APIs will bind words
loaded from strings into that context.

For now, the top level code (e.g. main()) runs in the user context.
Core code (built-in natives and actions/port-actions) all run in the
lib context.

Ideally, if each extension had its own isolated context to bind in,
then that would be used for extension natives.  This would let the
code written for an extension native see the usermode code it exports,
which would be ideal.  This sets up the groundwork for that change.