Restore BOUND? specialization removed to %redbol.reb
The BOUND? specialization was accidentally moved into %redbol.reb with
the BIND? specialization.  BOUND? was to be a LOGIC!-returning
chain/specialization that gave a true or false answer to whether
something is bound, thus properly be a funciton ending in "?"

Also removes the `<r3-legacy>` tag execution stub, now replaced by
the redbol.reb module.