Use Do_Va_Throws vs. lower-level eval routines in %a-lib.c
With the elimination of DO/NEXT, a concept for what "Do_XXX" means in
the API was established.  It's "runs all the way to completion" and
"defaults the output cell to void".

The meaning of "Do_Va_Throws" existed to do a variadic evaluation, but
it did not take a valist *pointer*.  Only Eval_Va_Core() took the
pointer, which routines like rebRun() or rebDid() had to use...because
they received an independent first parameter (a limitation of C, it
needs at least one fixed parameter to implement variadics).  So each
call to the variadic processing from the API was using a lower-level
API and passing in flags...testing an INDEXOR, etc.

Yet there were only two calls to this version of Do_Va_Throws, which
were in what would be considered "extension code"...so they should be
calling the libRebol API anyway.  This switched those two calls (in
the FREE routine for STRUCT! and the automatic compilation dispatcher
for the TCC build) to use rebElide().  Then it changes Do_Va_Throws()
to implement the needed repeated logic in the API...making it smaller
and clearer.

Includes a small simplification of port code, also noticed.