NULL native, use BLANK! instead of NULL as boot signal
In the attempt to divide the EVALUATE commit into smaller chunks, this
introduces NULL as a native...as a function with no body will now be
evaluating to #[void] by default, not to NULL.

Also, there had been use of `()` to generate NULLs as a signal at the
end of the boot sections when they would go through a DO, to make sure
they got all the way to the end.  Under the new rules, this would no
longer generate null but disappear...and using NULL would require a
word lookup, which convolutes the dependencies/binding in bootstrap.
For the sake of simplicity this just uses a BLANK! literal `_`.

Includes a change to IS_TRASH_DEBUG which recognizes that end cells
are not trash, even though they deliberately use trash as their type
bits...just to make sure you don't use those bits.