Eliminate DO_NEXT_MAY_THROW (dup of Eval_Array_At_Core)
This very old function had somehow evaded notice as nearly an exact
duplicate of the Eval_Array_At_Core() function.  It is unable to
specify a first element to initialize or flags for DO_TO_END, but
repeated most of the same code otherwise.  It included an automatic
subtraction of 1 from the `indexor` (since the evaluator prefetches),
which clients of Eval are expected to do themselves if they care (since
passing in DO_TO_END would imply they did not care).

There were only two relatively obscure uses: in the FFI struct code
and the poorly-defined/buggy DO feature in PARSE from R3-Alpha.

The FFI struct code--if it is to be written to the internal API--should
be using frame iteration with Push_Frame_At().  This would lock the
struct array so the pointers would be safe from corruption during the
enumeration via user code.  It would also allow usage of the same frame
for multiple evaluations.  However, such a rewrite would probably be
a waste of time as the breakdown of Rebol STRUCT! specifications into
foundational FFI parts is probably better done with Rebol code (e.g.
PARSE) in usermode, passing low-level by-products to the C.

As for the DO feature in PARSE, it was unfinished and undocumented, and
does not seem to be implemented in Red.

This commit makes the small change to use Eval_Array_At_Core() instead,
and removes the DO_NEXT_MAY_THROW routine.  Also removes an unwritten
function that was a stub for if an array of REBVAL[] would be passed
in that didn't live in a Rebol array, but there is no longer much of
a need for that given the REBVAL*-based API model.

Moves some documentation for Do_Branch_Throws() to Trello, and makes
that routine a little shorter.