DO/NEXT => EVALUATE, empty GROUP! transparency/barrier
This commit improves the interface by which a DO/NEXT is done.  Before,
each DO/NEXT would return the evaluative result (like DO) but allow
the updated position to be stored in a variable:

    >> do/next [1 + 2 10 + 20] 'pos
    == 3

    >> pos
    == [10 + 20]

This changes that to return the new position and *optionally* update
a variable with the output result.  Since this changes the contract
far from what DO offers, it gets the new name EVALUATE:

    >> evaluate/set [1 + 2 10 + 20] 'value
    == [10 + 20]

    >> value
    == 3

The optional update of the variable is key.  This means that in
successive evaluator steps, the output need not be overwritten.  This
deals with the problem of "residual invisibles", e.g.

    >> evaluate/set [1 + 2 elide (print "hi")] 'var
    == [elide (print "hi")]

    >> var
    == 3

    >> evaluate/set [elide (print "hi")] 'var
    ;-- null

    >> var
    == 3

With this dynamic in hand, it's possible to defer elided material so
that it runs along with what comes after it (if anything).  This
is preferable to running with the previous thing, which led to
undesirable behaviors like:

    >> any [1 + 2 elide (print "hi")]
    hi ;-- would have preferred NOT to run this
    == 3

Paralleling this mechanic in the core, the Eval_XXX routines are
separated out from the Do_XXX routines.  The DO-named routines assume
they will produce void from no input and always run to the end, while
the EVAL-named routines expect a non-trash input cell to be passed in
as the output, which it will only overwrite if it makes an evaluative
product.

While DO of an empty block synthesizes a void, this also changes the
behavior of GROUP!s to not synthesize any values...just to group them.
In interstitial positions they are allowed to evaporate entirely, but
when used as arguments they must come up with a value or they are
treated as an expression barrier.  Always erroring in this case is
another possibility which could be considered--but this is more lax to
test to see if it is useful.

(Note: This is a major change that significantly impacts the evaluator.
In an ideal world it would be tested and vetted more thoroughly, but
it seems a certain enough change to go ahead and push through and roll
with the results.)